+++ src/main/java/com/springboot/Application.java
@@ -1,4 +1,4 @@
-package com.springboot.employee.management;
+package com.springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
