+++ src/main/java/com/springboot/controller/EmployeeController.java
@@ -1,0 +1,39 @@
+package com.springboot.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import com.springboot.model.Employee;
+import com.springboot.service.EmployeeService;
+
+@Controller
+public class EmployeeController {
+	
+	@Autowired
+	private EmployeeService employeeService;
+	
+	@GetMapping("/")
+	public String viewHomePage(Model model) {
+		model.addAttribute("listEmployees", employeeService.getAllEmployees());
+		return "index";
+	}
+
+	@GetMapping("/showNewEmployeeForm")
+	public String showNewEmployeeForm(Model model) {
+		// create model attribute to bind form data
+		Employee employee = new Employee();
+		model.addAttribute("employee", employee);
+		return "new_employee";
+	}
+	
+	@PostMapping("/saveEmployee")
+	public String saveEmployee(@ModelAttribute("employee") Employee employee) {
+		// save employee to database
+		employeeService.saveEmployee(employee);
+		return "redirect:/";
+	}
+}
