--- src/main/java/com/springboot/empolyee/management/controller/EmployeeController.java
@@ -1,22 +1,0 @@
-package com.springboot.empolyee.management.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-import com.springboot.empolyee.management.service.EmployeeService;
-
-@Controller
-public class EmployeeController {
-	
-	@Autowired
-	private EmployeeService employeeService;
-	
-	@GetMapping("/")
-	public String viewHomePage(Model model) {
-		model.addAttribute("listEmployees", employeeService.getAllEmployees());
-		return "index";
-	}
-
-}
