--- src/main/java/com/springboot/empolyee/management/service/EmployeeService.java
@@ -1,10 +1,0 @@
-package com.springboot.empolyee.management.service;
-
-import java.util.List;
-
-import com.springboot.empolyee.management.model.Employee;
-
-public interface EmployeeService {
-
-	List<Employee> getAllEmployees();
-}
