+++ src/main/java/com/springboot/repository/EmployeeRepository.java
@@ -1,9 +1,9 @@
-package com.springboot.empolyee.management.repository;
+package com.springboot.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.springboot.empolyee.management.model.Employee;
+import com.springboot.model.Employee;
 
 
 @Repository
