+++ src/main/java/com/springboot/service/EmployeeService.java
@@ -1,0 +1,11 @@
+package com.springboot.service;
+
+import java.util.List;
+
+import com.springboot.model.Employee;
+
+public interface EmployeeService {
+
+	List<Employee> getAllEmployees();
+	void saveEmployee(Employee employee);
+}
