+++ src/main/java/com/springboot/service/EmployeeServiceImpl.java
@@ -1,12 +1,12 @@
-package com.springboot.empolyee.management.service;
+package com.springboot.service;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.springboot.empolyee.management.model.Employee;
-import com.springboot.empolyee.management.repository.EmployeeRepository;
+import com.springboot.model.Employee;
+import com.springboot.repository.EmployeeRepository;
 
 
 @Service
@@ -18,5 +18,10 @@ public class EmployeeServiceImpl implements EmployeeService {
 	@Override
 	public List<Employee> getAllEmployees() {
 		return employeeRepository.findAll();
+	}
+	
+	@Override
+	public void saveEmployee(Employee employee) {
+		this.employeeRepository.save(employee);
 	}
 }
