+++ src/main/resources/templates/index.html
@@ -3,29 +3,33 @@
 <head>
 <meta charset="ISO-8859-1">
 <title>Employee Management System</title>
+	<link rel="stylesheet"
+	href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
+	integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
+	crossorigin="anonymous">
 	
 </head>
 <body>
 
-	<div align="center">
+	<div class="container my-2">
 	<h1>Employees List</h1>
+	  
+<a th:href = "@{/showNewEmployeeForm}" class="btn btn-primary btn-sm mb-3"> Add Employee </a>
 	
-	<table border="1">
-	  <thead>
+	<table border="1" class = "table table-striped table-responsive-md">
+		<thead>
 	     <tr>
 	         <th> Employee First Name </th>
 	         <th> Employee Last Name </th>
 	         <th> Employee Email</th>
-	         </tr>
-	         </thead>
+	      </tr>
+	     </thead>
+	         
 	         <tbody>
 				<tr th:each="employee : ${listEmployees}">
 					<td th:text="${employee.firstName}"></td>
 					<td th:text="${employee.lastName}"></td>
 					<td th:text="${employee.email}"></td>
-					<td> <a th:href="@{/showFormForUpdate/{id}(id=${employee.id})}" class="btn btn-primary">Update</a>
-					    <a th:href="@{/deleteEmployee/{id}(id=${employee.id})}" class="btn btn-danger">Delete</a>
-					</td>
 				</tr>
 			</tbody>
 		</table>
